+++ CMakeLists.txt
@@ -11,6 +11,7 @@ add_library(${PROJECT_NAME}
         Network.cpp
         RTPacket.cpp
         RTProtocol.cpp
+        Settings.cpp
 )
 
 target_include_directories(${PROJECT_NAME} PUBLIC
