+++ Settings.cpp
@@ -1,0 +1,1 @@
+#include "Settings.h"
